#3686 comment: moved dev_config to local folder
