+++ api/tests/test_scripts/integration_tests/test_admin_tasks.py
@@ -25,7 +25,7 @@ import thiscovery_lib.utilities as utils
 
 
 from thiscovery_lib.dynamodb_utilities import Dynamodb
-from api.endpoints.common.dev_config import QUALTRICS_TEST_OBJECTS
+from api.local.dev_config import QUALTRICS_TEST_OBJECTS
 from api.local.admin_tasks.qualtrics_admin.convert_responses_to_contact_list_input_format import ResponsesToContactListConverter
 from api.local.admin_tasks.qualtrics_admin.republish_survey_with_user_specific_links import DistributionLinksGenerator
 from api.local.admin_tasks.task_setup.output_anon_project_specific_user_ids_for_test_group import ProcessManager as OutputAnonIds
